Fix AdaptiveSolver design issue
