+++ dynamiqs/sesolve/adaptive.py
@@ -1,6 +1,6 @@
 from torch import Tensor
 
-from ..solvers.ode.adaptive_solver import AdaptiveSolver
+from ..solvers.ode.adaptive_solver import AdaptiveSolver, DormandPrince5
 
 
 class SEAdaptive(AdaptiveSolver):
@@ -8,3 +8,7 @@ class SEAdaptive(AdaptiveSolver):
         """Compute dpsi / dt = -1j * H(psi) at time t."""
         # psi: (b_H, b_psi, n, 1) -> (b_H, b_psi, n, 1)
         return -1j * self.H(t) @ psi
+
+
+class SEDormandPrince5(SEAdaptive, DormandPrince5):
+    pass
